Initial docker-compose.yml and .env for running unbound + pi-hole
Contains the docker-compose.yml and .env files for running unbound and pi-hole in docker.  These have been tested locally on a Windows 10 machine.  As long as the OS or specific adapter have DNS servers as 127.0.0.1 and 1.1.1.1 (for example), this should download the images (initially using 1.1.1.1 to resolve).  Once the containers are running, 127.0.0.1 should resolve to this pi-hole container instance.